Move landing page to index.tsx
