+++ src/pages/index.tsx
@@ -1,101 +1,53 @@
+import { type NextPage } from "next"
 import { signIn, signOut, useSession } from "next-auth/react";
-import Head from "next/head";
-import Link from "next/link";
-import Image from "next/image";
-import { api } from "~/utils/api";
+import Image from "next/image"
+import Link from "next/link"
 
-export default function Home() {
-  const hello = api.example.hello.useQuery({ text: "from tRPC" });
-
-  console.error("welcome");
-
-  return (
-    <>
-      <Head>
-        <title>Create T3 App</title>
-        <meta name="description" content="Generated by create-t3-app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <main className="flex min-h-screen flex-col items-center justify-center bg-gradient-to-b from-[#2e026d] to-[#15162c]">
-        <div className="container flex flex-col items-center justify-center gap-12 px-4 py-16 ">
-          <h1 className="font-heading text-5xl font-extrabold tracking-tight text-white sm:text-[5rem]">
-            Create <span className="text-[hsl(280,100%,70%)]">T3</span> App
-          </h1>
-          <div className="grid grid-cols-1 gap-4 sm:grid-cols-2 md:gap-8">
-            <Link
-              className="flex max-w-xs flex-col gap-4 rounded-xl bg-white/10 p-4 text-white hover:bg-white/20"
-              href="https://create.t3.gg/en/usage/first-steps"
-              target="_blank"
-            >
-              <h3 className="font-heading text-2xl font-bold">First Steps →</h3>
-              <div className="font-body text-lg">
-                Just the basics - Everything you need to know to set up your
-                database and authentication.
-              </div>
-            </Link>
-            <Link
-              className="flex max-w-xs flex-col gap-4 rounded-xl bg-white/10 p-4 text-white hover:bg-white/20"
-              href="https://create.t3.gg/en/introduction"
-              target="_blank"
-            >
-              <h3 className="font-heading text-2xl font-bold">
-                Documentation →
-              </h3>
-              <div className="font-body text-lg">
-                Learn more about Create T3 App, the libraries it uses, and how
-                to deploy it.
-              </div>
-            </Link>
-          </div>
-          <div className="flex flex-col items-center gap-2">
-            <p className="font-body text-2xl text-white">
-              {hello.data ? hello.data.greeting : "Loading tRPC query..."}
-            </p>
-            <AuthShowcase />
-          </div>
-          {/* <QRScanner /> */}
-        </div>
-      </main>
-    </>
-  );
-}
-
-function AuthShowcase() {
+const Landing: NextPage = () => {
+  const links = [
+    {
+      href: "user/dashboard",
+      name: "Dashboard",
+    },
+    {
+      href: "leaderboard",
+      name: "Leaderboard",
+    },
+    {
+      href: "about-us",
+      name: "About Us",
+    },
+  ]
   const { data: sessionData } = useSession();
-
-  const { data: secretMessage } = api.example.getSecretMessage.useQuery(
-    undefined, // no input
-    { enabled: sessionData?.user !== undefined }
-  );
-
-  return (
-    <div className="flex flex-col items-center justify-center gap-4 font-body">
-      <p className="text-center text-2xl text-white">
-        {sessionData && <span>Logged in as {sessionData.user?.name}</span>}
-        {secretMessage && <span> - {secretMessage}</span>}
-      </p>
-      <p className="text-center text-2xl text-white">
-        {sessionData && `user id ${sessionData.user?.id}`}
-      </p>
-      <p className="text-center text-2xl text-white">
-        {sessionData && `email ${sessionData.user?.email}`}
-      </p>
-
-      {sessionData && (
-        <Image
-          alt="profile-picture"
-          src={sessionData.user?.image ? sessionData.user?.image : ""}
-          width={500}
-          height={500}
-          className="h-24 w-24 rounded-full"
-        />
-      )}
-      <button
-        className="rounded-full bg-white/10 px-10 py-3 font-semibold text-white no-underline transition hover:bg-white/20"
-        onClick={sessionData ? () => void signOut() : () => void signIn()}
-      >
-        {sessionData ? "Sign out" : "Sign in"}
-      </button>
+  
+  return (<>
+    <div className="w-full min-h-screen top-0 left-0 bg-[#FFE6EB]">
+      <div className="w-[320px] lg:w-[720px] min-h-screen lg:grid lg:grid-cols-2 m-auto">
+        <div className="relative w-[310px] h-[310px] py-8 lg:-top-8 m-auto">
+          <Image src="/landing-branch-fade.png" alt="" width={240} height={180} className="block absolute z-20 left-0 top-20" />
+          <Image src="/landing-branch-fade.png" alt="" width={240} height={180} className="block absolute -top-12 right-0 z-20 rotate-180" />
+          <div className="relative w-[300px] h-[300px] m-auto bg-gradient-to-b from-white to-[#FDB3C2] rounded-full">
+            <div className="relative w-full pl-12 py-28 z-30">
+              <h1 className="z-30 font-heading font-bold text-[#9F6C48] text-3xl">MegaLAN</h1>
+              <h2 className="z-30 font-heading font-bold text-brown text-4xl">Full Blossom</h2>
+            </div>
+          </div>
+        </div>
+        <div className="relative w-[320px] p-8 z-40 m-auto bg-gradient-to-b from-transparent via-20% via-[#FFE6EB] to-[#FFE6EB]">
+          {links.map((link) => (
+            <Link href={link.href} key={link.href} className="block my-8 p-[2px] bg-gradient-to-b from-[#FDB5C4] to-[#C58895] rounded-md">
+              <div className="p-2 rounded-md bg-[#FED7DF] font-heading text-pink text-center font-bold text-xl">{link.name}</div>
+            </Link>
+          ))}
+          <button className="block w-full my-8 p-[2px] bg-gradient-to-b from-[#FDB5C4] to-[#C58895] rounded-md">
+            <div onClick={sessionData ? () => void signOut() : () => void signIn()} className="p-2 rounded-md bg-[#FED7DF] font-heading text-[#EA5873] text-center font-bold text-xl">
+              {sessionData ? "Sign out" : "Sign in"}
+            </div>
+          </button>
+        </div>
+      </div>
     </div>
-  );
+  </>)
 }
+
+export default Landing
